Restructure version and main entrypoint logic
