+++ setup.py
@@ -1,6 +1,5 @@
 import os
 from setuptools import setup
-from _version import __version__
 
 """
 
@@ -22,6 +21,10 @@ Note: PyInstaller is used for generation of executables but not included in this
       some systems this may be blocked and therefore the directory of files is preferred.
 
 """
+
+# Imports the __version__ since the package references doesn't yet exist in the scope of setup.py. It opens the file
+# and interprets the code so the __version__ variable is populated, despite IDE warnings that it's undefined.
+exec (open('sqlite_dissect/_version.py').read())
 
 # The text of the README file
 README = open(os.path.join(os.path.dirname(__file__), 'README.md')).read()
