+++ sqlite_dissect/_version.py
