+++ sqlite_dissect/export/case_export.py
@@ -9,7 +9,7 @@ import uuid
 from datetime import datetime
 from os import path
 
-from _version import __version__
+from sqlite_dissect._version import __version__
 from sqlite_dissect.utilities import hash_file
